+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add net/http client integration ([#876](https://github.com/getsentry/sentry-go/pull/876))
+
 ## 0.29.0
 
 The Sentry SDK team is happy to announce the immediate availability of Sentry Go SDK v0.29.0.